refactor toolbox
